add
update code

+++ SMS-With-Maven/src/main/java/com/mdfaysalhossain/SMS/With/Maven/repository/ISyllabusRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISyllabusRepo extends JpaRepository<SyllabusModel, Integer> {
+public interface ISyllabusRepo extends JpaRepository<SyllabusModel, Long> {
